Unit tests were added to check if the ID or username already exists in the database.
