+++ src/test/java/com/mhm/bank/controller/AuthControllerTest.java
@@ -4,237 +4,171 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.mhm.bank.dto.UserInformation;
-import com.mhm.bank.dto.UserRegisteredEvent;
-import com.mhm.bank.repository.UserRepository;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.junit.jupiter.api.AfterEach;
+import com.mhm.bank.exception.UserAlreadyExistsException;
+import com.mhm.bank.service.AuthService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.kafka.support.serializer.JsonDeserializer;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.web.servlet.MockMvc;
-import org.testcontainers.containers.KafkaContainer;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
-import java.time.Duration;
 import java.time.LocalDate;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
 
-import static org.hamcrest.Matchers.containsString;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
-@SpringBootTest
-@Testcontainers
-@AutoConfigureMockMvc
-@ActiveProfiles("test")
+@ExtendWith(MockitoExtension.class)
 class AuthControllerTest {
-    @Autowired
-    private MockMvc mockMvc;
-    @Autowired
-    private ObjectMapper objectMapper;
-    @Autowired
+
+    @Mock
+    private AuthService authService;
+
+    @InjectMocks
     private AuthController authController;
-    private KafkaConsumer<String, UserRegisteredEvent> consumer;
-    @Autowired
-    private UserRepository userRepository;
 
-    @Container
-    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>(
-            "postgres:15-alpine"
-    ).withDatabaseName("test_db")
-            .withUsername("test")
-            .withPassword("test")
-            .withReuse(false);
-
-    @Container
-    static KafkaContainer kafka = new KafkaContainer(
-            DockerImageName.parse("confluentinc/cp-kafka:7.3.3")
-                    .asCompatibleSubstituteFor("confluentinc/cp-kafka")
-    );
-
-    @DynamicPropertySource
-    static void configureProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", postgres::getJdbcUrl);
-        registry.add("spring.datasource.username", postgres::getUsername);
-        registry.add("spring.datasource.password", postgres::getPassword);
-
-        registry.add("spring.kafka.bootstrap-servers", kafka::getBootstrapServers);
-        registry.add("spring.kafka.producer.key-serializer", () -> "org.apache.kafka.common.serialization.StringSerializer");
-        registry.add("spring.kafka.producer.value-serializer", () -> "org.springframework.kafka.support.serializer.JsonSerializer");
-
-    }
+    private ObjectMapper objectMapper;
 
     @BeforeEach
     void setUp() {
+        objectMapper = new ObjectMapper();
         objectMapper.registerModule(new JavaTimeModule());
         objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-
-        // Configure the Kafka consumer for testing
-        setupKafkaConsumer();
-
-    }
-
-    @AfterEach
-    void tearDown() {
-        if (consumer != null) {
-            consumer.close();
-        }
-
-        if (!postgres.isRunning()) {
-            postgres.start();
-        }
-
-        if (!kafka.isRunning()) {
-            kafka.start();
-            setupKafkaConsumer();
-        }
-    }
-
-    private void setupKafkaConsumer() {
-        Map<String, Object> consumerProps = new HashMap<>();
-        consumerProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafka.getBootstrapServers());
-        consumerProps.put(ConsumerConfig.GROUP_ID_CONFIG, "test-group");
-        consumerProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
-        consumerProps.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-        consumerProps.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, JsonDeserializer.class);
-        consumerProps.put(JsonDeserializer.TRUSTED_PACKAGES, "*");
-        consumerProps.put(JsonDeserializer.TYPE_MAPPINGS, "userRegisteredEvent:com.mhm.bank.dto.UserRegisteredEvent");
-
-        if (consumer != null) {
-            consumer.close();
-        }
-
-        consumer = new KafkaConsumer<>(consumerProps, new StringDeserializer(),
-                new JsonDeserializer<>(UserRegisteredEvent.class, false));
-        consumer.subscribe(Collections.singletonList("user-registered"));
     }
 
     @Test
-    void registerUser_shouldReturnBadRequest_whenInputIsInvalid() throws Exception {
-        UserInformation userInformation = new UserInformation(
-                null,
-                null,
-                "",
-                "John",
-                "Doe",
-                "123 Main St",
-                "invalid-email",
-                null,
-                "123456789"
-        );
-
-        mockMvc.perform(post("/api/auth/register")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(userInformation)))
-                .andExpect(status().isBadRequest());
-    }
-
-    @Test
-    void registerUser_shouldSuccessfullyCreateUser_whenInputIsValid() throws Exception {
-        UserInformation validUser = new UserInformation(
+    void registerUser_shouldReturnCreated_whenUserRegistrationIsSuccessful() throws Exception {
+        UserInformation userInfo = new UserInformation(
                 "12345678",
-                "username123",
+                "testuser",
                 "Password123!",
                 "John",
                 "Doe",
                 "123 Main St",
-                "john.doe@example.com",
-                LocalDate.parse("1990-01-01"),
+                "john@example.com",
+                LocalDate.of(1990, 1, 1),
                 "123456789"
         );
 
-        mockMvc.perform(post("/api/auth/register")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(validUser)))
-                .andExpect(status().isCreated())
-                .andExpect(content().string("User with ID 12345678 has been added"));
+        when(authService.registerUser(any(UserInformation.class)))
+                .thenReturn("User with ID 12345678 has been added");
 
-        ConsumerRecords<String, UserRegisteredEvent> records = consumer.poll(Duration.ofSeconds(10));
-        assertFalse(records.isEmpty(), "No messages were received in Kafka");
+        ResponseEntity<String> response = authController.registerUser(userInfo);
 
-        ConsumerRecord<String, UserRegisteredEvent> recordConsumer = records.iterator().next();
-        UserRegisteredEvent event = recordConsumer.value();
-
-        assertNotNull(event, "Kafka event cannot be null");
-        assertEquals(validUser.idCard(), event.userId(), "User ID does not match");
-        assertEquals(validUser.username(), event.username(), "Username does not match");
-        assertEquals(validUser.email(), event.email(), "Email does not match");
+        assertEquals(HttpStatus.CREATED, response.getStatusCode());
+        assertEquals("User with ID 12345678 has been added", response.getBody());
+        verify(authService).registerUser(userInfo);
     }
 
     @Test
-    void registerUser_shouldFailWithInternalServerError_whenDatabaseIsDown() throws Exception {
-        postgres.stop();
-
-        UserInformation validUser = new UserInformation(
-                "234567890",
-                "username234",
-                "Password234!",
+    void registerUser_shouldThrowException_whenUserAlreadyExists() throws Exception {
+        UserInformation userInfo = new UserInformation(
+                "12345678",
+                "testuser",
+                "Password123!",
                 "John",
                 "Doe",
                 "123 Main St",
-                "john.doe@example.com",
-                LocalDate.parse("1990-01-01"),
+                "john@example.com",
+                LocalDate.of(1990, 1, 1),
                 "123456789"
         );
 
-        mockMvc.perform(post("/api/auth/register")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(validUser)))
-                .andExpect(status().isInternalServerError())
-                .andExpect(jsonPath("$.errorCode").value("DATABASE_ERROR"))
-                .andExpect(jsonPath("$.message").value(containsString("Error accessing the database")));
+        when(authService.registerUser(any(UserInformation.class)))
+                .thenThrow(new UserAlreadyExistsException("User already exists"));
 
-        ConsumerRecords<String, UserRegisteredEvent> records = consumer.poll(Duration.ofSeconds(5));
-        assertTrue(records.isEmpty(), "No messages should have been received in Kafka");
-
-        postgres.start();
+        assertThrows(UserAlreadyExistsException.class, () -> authController.registerUser(userInfo));
+        verify(authService).registerUser(userInfo);
     }
 
     @Test
-    void registerUser_shouldFailWithInternalServerError_whenKafkaIsDown() throws Exception {
-        kafka.stop();
-
-        UserInformation validUser = new UserInformation(
-                "3456789012",
-                "username345",
-                "Password345!",
+    void registerUser_shouldThrowException_whenEmailIsInvalid() throws Exception {
+        UserInformation userInfo = new UserInformation(
+                "12345678",
+                "testuser",
+                "Password123!",
                 "John",
                 "Doe",
                 "123 Main St",
-                "john.doe@example.com",
-                LocalDate.parse("1990-01-01"),
+                "invalid.email",
+                LocalDate.of(1990, 1, 1),
                 "123456789"
         );
 
-        mockMvc.perform(post("/api/auth/register")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(validUser)))
-                .andExpect(status().isInternalServerError())
-                .andExpect(jsonPath("$.errorCode").value("KAFKA_ERROR"))
-                .andExpect(jsonPath("$.message").value(containsString("Error with message broker")));
+        when(authService.registerUser(any(UserInformation.class)))
+                .thenThrow(new IllegalArgumentException("Invalid email format"));
 
-        assertFalse(userRepository.existsById(validUser.idCard()),
-                "User should not exist in database when Kafka is down");
-
-        kafka.start();
-        setupKafkaConsumer();
+        assertThrows(IllegalArgumentException.class, () -> authController.registerUser(userInfo));
+        verify(authService).registerUser(userInfo);
     }
+
+    @Test
+    void registerUser_shouldThrowException_whenUserIsUnderAge() throws Exception {
+        UserInformation userInfo = new UserInformation(
+                "12345678",
+                "testuser",
+                "Password123!",
+                "John",
+                "Doe",
+                "123 Main St",
+                "john@example.com",
+                LocalDate.now().minusYears(17),
+                "123456789"
+        );
+
+        when(authService.registerUser(any(UserInformation.class)))
+                .thenThrow(new IllegalArgumentException("User must be at least 18 years old"));
+
+        assertThrows(IllegalArgumentException.class, () -> authController.registerUser(userInfo));
+        verify(authService).registerUser(userInfo);
+    }
+
+    @Test
+    void registerUser_shouldThrowException_whenIdExists() throws Exception {
+        UserInformation userInfo = new UserInformation(
+                "12345678",
+                "newusername",
+                "Password123!",
+                "Jane",
+                "Smith",
+                "456 Oak St",
+                "jane@example.com",
+                LocalDate.of(1992, 1, 1),
+                "987654321"
+        );
+
+        when(authService.registerUser(any(UserInformation.class)))
+                .thenThrow(new UserAlreadyExistsException("User ID already registered"));
+
+        assertThrows(UserAlreadyExistsException.class, () -> authController.registerUser(userInfo));
+        verify(authService).registerUser(userInfo);
+    }
+
+    @Test
+    void registerUser_shouldThrowException_whenUsernameExistsWithDifferentData() throws Exception {
+        UserInformation userInfo = new UserInformation(
+                "87654321",
+                "testuser",
+                "Password123!",
+                "Jane",
+                "Smith",
+                "456 Oak St",
+                "jane@example.com",
+                LocalDate.of(1992, 1, 1),
+                "987654321"
+        );
+
+        when(authService.registerUser(any(UserInformation.class)))
+                .thenThrow(new UserAlreadyExistsException("Username already taken"));
+
+        assertThrows(UserAlreadyExistsException.class, () -> authController.registerUser(userInfo));
+        verify(authService).registerUser(userInfo);
+    }
+
 
 }
